fixed the fps not capping issue

+++ app/src/main/java/com/graphics/glcanvas/engine/Updatable.kt
@@ -3,6 +3,6 @@ package com.graphics.glcanvas.engine
 interface Updatable {
 
     fun draw()
-    fun update(delta:Float)
+    fun update(delta:Long)
     fun prepare()
 }
